PM / hibernate: Reserve swsusp key and earse footprints
Add handler to parse the setup data that carrying swsusp key, it
reserves swsusp key by memblock then copies key to a allocated page
in later initcall stage.

And for earsing footbprints, the codes in this patch remove setup
data that carried swsusp key, and clean the memory space that
reserved by memblock.

Signed-off-by: Lee, Chun-Yi <jlee@suse.com>