PM / hibernate: a option to request that snapshot image must be validly signed
This kernel option is similar with the option for kernel module signature
verification. When the option is unselected, kernel will be tainted by
restoring from a unsigned snapshot image or the image was signed by a
wrong key.

When the option is selected, kernel will refuse the system to be resumed
if the signature verification did not pass. The hibernation resume
process will be stopped and whole system just boots as normal.

Signed-off-by: Lee, Chun-Yi <jlee@suse.com>