PM / hibernate: create snapshot keys handler and using EFI key
This patch adds a snapshot keys handler as a helper to use the
key retention service api to create the session key for snapshot
image encryption and authentication.

The handler uses EFI key as the snapshot key, and the encription
key and authentication key are derived from the snapshot key. The
encrypted key can also be a snapshot key, the logic will be added
to handler later.

The sanpshot key is a session key for one time usage. Either EFI
key or encrypted key can encrypt the key data to be a encrypted
blob. Kernel stores the encrypted blob to the snapshot image header
for hibernation resume.