PM / hibernate: Generate and verify signature for snapshot image
When producing memory snapshot image, hibernation uses HMAC-SHA512
with encrypted secret key to calculate the hash for all data pages
in image. The hash result will be kept in the snapshot header as the
image signature. Before hibernation restores system, kernel executes
HMAC-SHA512 again and compares the result with the signature in header
to verify the integrity of snapshot image.

If the verification failed, the resume process will be stopped. The
snapshot image will be discarded and system will boot as normal.