PM / hibernate: Erase the snapshot master key in snapshot pages
If the encryption key be guessed then the snapshot master key can
also be grabbed from snapshot image. Which means that the authentication
key can also be calculated. So kernel erases master key in snapshot
pages.

Because the master key in image kernel be erased, kernel uses the
trampoline page to forward snapshot master key to image kernel.

Signed-off-by: Lee, Chun-Yi <jlee@suse.com>